[fix] exports from module.js
Fix #666
