+++ src/index.js
@@ -1,136 +1,2 @@
-import {
-  // top-level API
-  findNodeHandle,
-  hydrate,
-  render,
-  unmountComponentAtNode,
-
-  // modules
-  createElement,
-  NativeModules,
-  processColor,
-
-  // APIs
-  Animated,
-  AppRegistry,
-  AppState,
-  AsyncStorage,
-  BackAndroid,
-  Clipboard,
-  Dimensions,
-  Easing,
-  I18nManager,
-  InteractionManager,
-  Keyboard,
-  Linking,
-  NetInfo,
-  PanResponder,
-  PixelRatio,
-  Platform,
-  StyleSheet,
-  UIManager,
-  Vibration,
-
-  // components
-  ActivityIndicator,
-  Button,
-  FlatList,
-  Image,
-  KeyboardAvoidingView,
-  ListView,
-  Modal,
-  Picker,
-  ProgressBar,
-  RefreshControl,
-  ScrollView,
-  SectionList,
-  Slider,
-  StatusBar,
-  Switch,
-  Text,
-  TextInput,
-  Touchable,
-  TouchableHighlight,
-  TouchableNativeFeedback,
-  TouchableOpacity,
-  TouchableWithoutFeedback,
-  View,
-  VirtualizedList,
-
-  // propTypes
-  ColorPropType,
-  EdgeInsetsPropType,
-  PointPropType,
-  TextPropTypes,
-  ViewPropTypes
-} from './module';
-
-const ReactNative = {
-  // top-level API
-  findNodeHandle,
-  hydrate,
-  render,
-  unmountComponentAtNode,
-
-  // modules
-  createElement,
-  NativeModules,
-  processColor,
-
-  // APIs
-  Animated,
-  AppRegistry,
-  AppState,
-  AsyncStorage,
-  BackAndroid,
-  BackHandler: BackAndroid,
-  Clipboard,
-  Dimensions,
-  Easing,
-  I18nManager,
-  InteractionManager,
-  Keyboard,
-  Linking,
-  NetInfo,
-  PanResponder,
-  PixelRatio,
-  Platform,
-  StyleSheet,
-  UIManager,
-  Vibration,
-
-  // components
-  ActivityIndicator,
-  Button,
-  FlatList,
-  Image,
-  KeyboardAvoidingView,
-  ListView,
-  Modal,
-  Picker,
-  ProgressBar,
-  RefreshControl,
-  ScrollView,
-  SectionList,
-  Slider,
-  StatusBar,
-  Switch,
-  Text,
-  TextInput,
-  Touchable,
-  TouchableHighlight,
-  TouchableNativeFeedback,
-  TouchableOpacity,
-  TouchableWithoutFeedback,
-  View,
-  VirtualizedList,
-
-  // propTypes
-  ColorPropType,
-  EdgeInsetsPropType,
-  PointPropType,
-  TextPropTypes,
-  ViewPropTypes
-};
-
+import ReactNative from './module';
 module.exports = ReactNative;
